+++ src/main/java/com/example/demo/dto/OrderLineResponseDto.java
@@ -2,13 +2,11 @@ package com.example.demo.dto;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 
 @Data
 public class OrderLineResponseDto {
     private String productName;
     private String productReference;
-    private BigDecimal quantity;
-    private BigDecimal price;
+    private Integer quantity;
+    private float price;
 }
